requirements.txt
markitdown>=0.0.2
markdown>=3.4.4
python-docx>=0.8.11
python-pptx>=0.6.21
openpyxl>=3.1.2
html2docx>=0.6.0
